+++ src/stories/InputField.stories.tsx
@@ -27,6 +27,8 @@ const mockHandleInput = action('handleInput');
 
 export const Default: Story = {
   args: {
+    title: '기본 입력 필드 제목',
+    subtitle: '기본 입력 필드 부제목',
     inputTypes: {
       inputLabel: '기본 입력 필드',
       inputInfo: Array.from({ length: 1 }, (_, index) => ({
@@ -43,6 +45,8 @@ export const Default: Story = {
 
 export const CARD_NUMBER_FIELD: Story = {
   args: {
+    title: '결제할 카드 번호를 입력해 주세요',
+    subtitle: '본인 명의의 카드만 결제 가능합니다.',
     inputTypes: INPUT_TYPE_CATEGORIES.CARD_NUMBER,
     handleInput: mockHandleInput,
   },
@@ -50,6 +54,8 @@ export const CARD_NUMBER_FIELD: Story = {
 
 export const EXPIRY_DATE_FIELD: Story = {
   args: {
+    title: '카드 유효기간을 입력해 주세요',
+    subtitle: '월/년도(MMYY)를 순서대로 입력해 주세요.',
     inputTypes: INPUT_TYPE_CATEGORIES.EXPIRY_DATE,
     handleInput: mockHandleInput,
   },
@@ -57,6 +63,7 @@ export const EXPIRY_DATE_FIELD: Story = {
 
 export const USER_NAME_FIELD: Story = {
   args: {
+    title: '카드 소유자 이름을 입력해 주세요',
     inputTypes: INPUT_TYPE_CATEGORIES.USER_NAME,
     handleInput: mockHandleInput,
   },
